% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSetClass.R
\name{annotation<-}
\alias{annotation<-}
\title{annotation<- Slot Setter}
\usage{
annotation(object, ...) <- value
}
\arguments{
\item{object}{A \code{RadioSet}}

\item{...}{A \code{list} to allow definition of new parameters on this generic}

\item{value}{A \code{list} of annotations to add to the annotatiosn slot of
an rSet}
}
\value{
A copy of the \code{RadioSet} with the updated annotation slot
}
\description{
annotation<- Slot Setter
}
\examples{
data(clevelandSmall)
annotation(clevelandSmall) <- annotation(clevelandSmall)

}
