% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRSet.R
\name{availableRSets}
\alias{availableRSets}
\title{Return a table of RadioSets available for download}
\usage{
availableRSets(
  saveDir = tempdir(),
  fileName = "availableRadioSets.csv",
  verbose = TRUE
)
}
\arguments{
\item{saveDir}{\code{character} Directory to save the table of rSets}

\item{fileName}{\code{character} The filename for the table of rSets}

\item{verbose}{\code{bool} Should status messages be printed during download.}
}
\value{
A data.frame with details about the available RadioSet objects
}
\description{
The function fetches a table of all RadioSets available for download from
the PharmacoGx server. The table includes the names of the PharamcoSet, the
types of data available in the object, and the date of last update.
}
\examples{
if (interactive()){
availableRSets()
}

}
