% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Cleveland_small}
\alias{Cleveland_small}
\title{Cleaveland_mut RadioSet subsetted old}
\format{
RadioSet object
}
\usage{
data(Cleveland_small)
}
\description{
This is the old version where meolecularProfiles are ExpressionSets instead
  of SummarizedExperiments
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect
  small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
