% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{clevelandSmall}
\alias{clevelandSmall}
\title{Cleaveland_mut RadioSet subsetted}
\format{
RadioSet object
}
\usage{
data(clevelandSmall)
}
\description{
Documentation for this dataset will be added at a later date. For now I just
need this package to pass the CRAN checks! This dataset powers the exampe
usage in the roxygen2 documentation for CoreGx.
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect
  small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
