% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAUC.R
\name{computeAUC}
\alias{computeAUC}
\title{computeAUC: computes AUC}
\usage{
computeAUC(alpha, beta, lower, upper, SF_as_log = TRUE)
}
\arguments{
\item{alpha}{parameter in equation y = exp(-alpha * x - beta * x ^ 2)}

\item{beta}{parameter in equation y = exp(-alpha * x - beta * x ^ 2)}

\item{lower}{lower bound of dose region to compute AUC over}

\item{upper}{upper bound of dose region to compute AUC over}

\item{SF_as_log}{should AUC be computed with log10(survival fraction) instead of survival fraction? Defaults to FALSE.}
}
\description{
This function computes the area under a dose-response curve of the form survival fraction SF = exp(-alpha * D - beta * D ^ 2).
}
