% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeSF2.R
\name{computeSF2}
\alias{computeSF2}
\title{Compute SF2}
\usage{
computeSF2(alpha, beta, SF_as_log = TRUE)
}
\arguments{
\item{alpha}{parameter in equation y = exp(-alpha * x - beta * x ^ 2)}

\item{beta}{parameter in equation y = exp(-alpha * x - beta * x ^ 2)}

\item{SF_as_log}{should log10(SF2) be returned instead of SF2? Defaults to TRUE.}
}
\description{
This function computes the survival fraction after administering 2 units of radiation, given alpha and beta in the equation SF = exp(-alpha * D - beta * D ^ 2).
}
