% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation-methods.R
\name{curation<-}
\alias{curation<-}
\title{curation<- Slot Setter}
\usage{
curation(object, ...) <- value
}
\arguments{
\item{object}{A \code{RadioSet}}

\item{...}{A \code{list} to allow definition of new parameters on this generic}

\item{value}{A \code{list} of curations for the cell and tissues types in the
rSet object}
}
\value{
A copy of the \code{RadioSet} with the updated curation slot
}
\description{
curation<- Slot Setter
}
\examples{
data(clevelandSmall)
curation(clevelandSmall) <- curation(clevelandSmall)

}
