% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curation-methods.R
\name{curation}
\alias{curation}
\title{curation Slot Getter}
\usage{
curation(object, ...)
}
\arguments{
\item{object}{A \code{RadioSet}}

\item{...}{A \code{list} to allow definition of new parameters on this generic}
}
\value{
A \code{list} of unique cell and tissue identifiers to check validity
  of an rSet
}
\description{
curation Slot Getter
}
\examples{
data(clevelandSmall)
curation(clevelandSmall)

}
