% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetType-methods.R
\name{datasetType<-}
\alias{datasetType<-}
\title{datasetType<- Replacement Generic}
\usage{
datasetType(object) <- value
}
\arguments{
\item{object}{A \code{RadioSet} from which to retrieve the dataset type}

\item{value}{A \code{character} vector containing the dataset type}
}
\value{
A \code{character} vector containing the dataset tpye
}
\description{
A generic for updating the dataset type of a RadioSet object
}
\examples{
data(clevelandSmall)
datasetType(clevelandSmall)

}
