% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlities.R
\name{.convertRsetMolecularProfilesToSE}
\alias{.convertRsetMolecularProfilesToSE}
\title{rSet molecularProfiles from ESets to SEs}
\usage{
.convertRsetMolecularProfilesToSE(rSet)
}
\description{
Converts all ExpressionSet objects within the molecularProfiles slot of a
}
