% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlities.R
\name{.validaterSetMolecularProfilesToSEConversion}
\alias{.validaterSetMolecularProfilesToSEConversion}
\title{Validate rSet molecularProfiles Conversion}
\usage{
.validaterSetMolecularProfilesToSEConversion(rSet_old, rSet_new)
}
\arguments{
\item{rSet_old}{\code{S4} a rSet containing molecularProfiles as ExpressionSets}

\item{rSet_new}{\code{S4} a rSet containing molecularProfiles as SummarizedExperiments}
}
\value{
\code{message} Any slots which are not the same
}
\description{
Checks that all the information contained in an ExpressionSet molecularProfile
  was successfully tranferred to the SummarizedExperiment molecularProfile
}
\keyword{internal}
