% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneDrugSensitivity.R
\name{geneDrugSensitivity}
\alias{geneDrugSensitivity}
\title{Calcualte the gene drug sensitivity score}
\usage{
geneDrugSensitivity(
  x,
  type,
  batch,
  drugpheno,
  interaction.typexgene = FALSE,
  model = FALSE,
  standardize = c("SD", "rescale", "none"),
  verbose = FALSE
)
}
\arguments{
\item{x}{\code{numeric} vector of gene expression values}

\item{type}{\code{factor} vector specifying the cell lines or type types}

\item{batch}{\code{factor} vector of factors specifying the batch}

\item{drugpheno}{\code{numeric} vector of drug sensitivity values (e.g.,
IC50 or AUC)}

\item{interaction.typexgene}{\code{logical} Should interaction between gene
expression and cell/type type be computed? Default set to FALSE}

\item{model}{Should the full linear model be returned? Default set to FALSE}

\item{standardize}{\code{character} One of 'SD', 'rescale' or 'none'}

\item{verbose}{\code{logical} Should informative messages be written to
the console?}
}
\value{
A \code{numeric} vector reporting the effect size (estimate of the
  coefficient of drug concentration), standard error (se), sample size (n),
  t statistic, and F statistics and its corresponding p-value
}
\description{
Calcualte the gene drug sensitivity score
}
