% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{mDataNames}
\alias{mDataNames}
\title{mDataNames}
\usage{
mDataNames(rSet)
}
\arguments{
\item{rSet}{PharamcoSet object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the RadioSet.
}
\examples{
data(CCLEsmall)
mDataNames(CCLEsmall)

}
