% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationNames<-}
\alias{radiationNames<-}
\title{radiationNames<- Generic}
\usage{
radiationNames(object) <- value
}
\arguments{
\item{object}{The \code{RadioSet} to update}

\item{value}{A \code{character} vector of the new radiation names}
}
\value{
Updated \code{RadioSet}
}
\description{
A generic for the radiationNames replacement method
}
\examples{
data(CCLEsmall)
radiationNames(CCLEsmall) <- radiationNames(CCLEsmall)

}
