% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationNames}
\alias{radiationNames}
\title{radiationNames Generic}
\usage{
radiationNames(rSet)
}
\arguments{
\item{rSet}{The \code{RadioSet} to return radiation names from}
}
\value{
A vector of the radiation names used in the RadioSet
}
\description{
A generic for the radiationNames method
}
\examples{
data(CCLEsmall)
radiationNames(CCLEsmall)

}
