% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSetClass.R
\docType{class}
\name{RadioSet-class}
\alias{RadioSet-class}
\alias{.RadioSet}
\alias{cellInfo,RadioSet-method}
\alias{cellInfo<-,RadioSet,data.frame-method}
\alias{radiationInfo,RadioSet-method}
\alias{radiationInfo<-,RadioSet,data.frame-method}
\alias{phenoInfo,RadioSet,character-method}
\alias{phenoInfo<-,RadioSet,character,data.frame-method}
\alias{molecularProfiles,RadioSet,character-method}
\alias{molecularProfiles<-,RadioSet,character,character,matrix-method}
\alias{molecularProfiles<-,RadioSet,character,missing,matrix-method}
\alias{molecularProfilesSlot,RadioSet-method}
\alias{molecularProfilesSlot<-,RadioSet,ANY-method}
\alias{featureInfo,RadioSet,character-method}
\alias{featureInfo<-,RadioSet,character,data.frame-method}
\alias{sensitivityInfo,RadioSet-method}
\alias{sensitivityRaw,RadioSet-method}
\alias{sensitivityRaw<-,RadioSet,array-method}
\alias{sensitivitySlot,RadioSet-method}
\alias{sensitivitySlot<-,RadioSet,list-method}
\alias{sensitivityInfo<-,RadioSet,DataFrame-method}
\alias{sensitivityProfiles,RadioSet-method}
\alias{sensitivityProfiles<-,RadioSet,data.frame-method}
\alias{sensitivityProfiles<-,RadioSet,matrix-method}
\alias{sensitivityMeasures,RadioSet-method}
\alias{radiationTypes,RadioSet-method}
\alias{radiationTypes<-,RadioSet,character-method}
\alias{cellNames,RadioSet-method}
\alias{cellNames<-,RadioSet,character-method}
\alias{fNames,RadioSet,character-method}
\alias{fNames<-,RadioSet,character,character-method}
\alias{dateCreated,RadioSet-method}
\alias{datasetType,RadioSet-method}
\alias{datasetType<-,RadioSet-method}
\alias{name,RadioSet-method}
\alias{pertNumber,RadioSet-method}
\alias{sensNumber,RadioSet-method}
\alias{pertNumber<-,RadioSet,array-method}
\alias{sensNumber<-,RadioSet,matrix-method}
\title{A Class to Contain RadioGenomic datasets together with their curations}
\usage{
\S4method{cellInfo}{RadioSet}(object)

\S4method{cellInfo}{RadioSet,data.frame}(object) <- value

\S4method{radiationInfo}{RadioSet}(object)

\S4method{radiationInfo}{RadioSet,data.frame}(object) <- value

\S4method{phenoInfo}{RadioSet,character}(object, mDataType)

\S4method{phenoInfo}{RadioSet,character,data.frame}(object, mDataType) <- value

\S4method{molecularProfiles}{RadioSet,character}(object, mDataType)

\S4method{molecularProfiles}{RadioSet,character,character,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfiles}{RadioSet,character,missing,matrix}(object, mDataType, assay) <- value

\S4method{molecularProfilesSlot}{RadioSet}(object)

\S4method{molecularProfilesSlot}{RadioSet,ANY}(object) <- value

\S4method{featureInfo}{RadioSet,character}(object, mDataType)

\S4method{featureInfo}{RadioSet,character,data.frame}(object, mDataType) <- value

\S4method{sensitivityInfo}{RadioSet}(object)

\S4method{sensitivityRaw}{RadioSet}(object)

\S4method{sensitivityRaw}{RadioSet,array}(object) <- value

\S4method{sensitivitySlot}{RadioSet}(object)

\S4method{sensitivitySlot}{RadioSet,list}(object) <- value

\S4method{sensitivityInfo}{RadioSet,DataFrame}(object) <- value

\S4method{sensitivityProfiles}{RadioSet}(object)

\S4method{sensitivityProfiles}{RadioSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{RadioSet,matrix}(object) <- value

\S4method{sensitivityMeasures}{RadioSet}(object)

\S4method{radiationTypes}{RadioSet}(object)

\S4method{radiationTypes}{RadioSet,character}(object) <- value

\S4method{cellNames}{RadioSet}(object)

\S4method{cellNames}{RadioSet,character}(object) <- value

\S4method{fNames}{RadioSet,character}(object, mDataType)

\S4method{fNames}{RadioSet,character,character}(object, mDataType) <- value

\S4method{dateCreated}{RadioSet}(object)

\S4method{datasetType}{RadioSet}(object)

\S4method{datasetType}{RadioSet}(object) <- value

\S4method{name}{RadioSet}(object)

\S4method{pertNumber}{RadioSet}(object)

\S4method{sensNumber}{RadioSet}(object)

\S4method{pertNumber}{RadioSet,array}(object) <- value

\S4method{sensNumber}{RadioSet,matrix}(object) <- value
}
\arguments{
\item{object}{A \code{RadioSet}}

\item{value}{A new \code{matrix} with the number of sensitivity
experiments per drug and cell line}

\item{mDataType}{The molecular data type to update}

\item{assay}{\code{character} Name or index of the assay data to return}
}
\value{
An object of the RadioSet class

a \code{data.frame} with the cell annotations

Updated \code{RadioSet}

a \code{data.frame} with the experiment info

The updated \code{RadioSet}

a \code{data.frame} with the experiment info

Updated \code{RadioSet}

A \code{list} containing the molecularProfiles from a cSet

A copy of the \code{CoreSet} with the molecularProfiles slot updated

A \code{DataFrame} containing the feature information

Updated \code{RadioSet}

a \code{DataFrame} with the experiment info

Updated \code{RadioSet}

a \code{data.frame} with the experiment info

Updated \code{RadioSet}

A \code{character} vector of all the available sensitivity measures

A \code{vector} of the cell line names in the RadioSet

Updated \code{RadioSet}

A \code{character} vector of the feature names

Updated \code{RadioSet}

The date the RadioSet was created

The name of the RadioSet

A 3D \code{array} with the number of perturbation experiments per
  radiation type and cell line, and data type

A \code{data.frame} with the number of sensitivity experiments per
  radiation type and cell line

The updated \code{RadioSet}

The updated \code{RadioSet}
}
\description{
The RadioSet (RSet) class was developed to contain and organise large
RadioGenomic datasets, and aid in their metanalysis. It was designed
primarily to allow bioinformaticians and biologists to work with data at the
level of genes and cell lines, providing a more naturally intuitive
interface and simplifying analyses between several datasets. As such, it was
designed to be flexible enough to hold datasets of two different natures
while providing a common interface. The class can accomidate datasets
containing both radiation dose response data, as well as datasets contaning
genetic profiles of cell lines pre and post treatement with compounds, known
respecitively as sensitivity and perturbation datasets.

Get cell line information from a RadioSet object

Set cell line annotations for a RadioSet object

Getter method for the phenotype information of a given molecular data type

Setter method for the phenotype information of a given molecular data type

Getter method for the molecular profile data of the given molecular data
  type

Setter method for the molecular profile data of the given molecular data
  type

Getter method for the feature data of the given molecular data
  type

Setter method for the feature data of the given molecular data
  type

Get the sensitivity information for the cell lines in the RadioSet

Set the sensitivity information for the cell lines in the RadioSet

Get the sensitivity values for the cell lines in the RadioSet

Set the sensitivity information for the cell lines in the RadioSet

Get the types of sensitivity measurements available in a RadioSet

Get the names of all cell lines in a RadioSet object

Set the names of all cell lines in a RadioSet object

Return the feature names for the specified molecular data type

Setter for the feature names of a \code{SummarizedExperiment} in the
  molecularProfiles slot

Get the date a RadioSet object was created

Get the name of a RadioSet

Get a summary of available perturbation experiments

Get a summary of the available sensitivity experiments in the RadioSet

Get the summary of the available sensitivity experiments in the RadioSet

Set the summary of the available sensitivity experiments in the RadioSet
}
\section{Methods (by generic)}{
\itemize{
\item \code{cellInfo}: Returns the annotations for all the cell lines tested
on in the RadioSet

\item \code{cellInfo<-}: Update the cell line annotations

\item \code{radiationInfo}: Returns the annotations for all the radiations tested
in the RadioSet

\item \code{radiationInfo<-}: Update the radiation annotations

\item \code{phenoInfo}: Return the experiment info from the given type of
molecular data in RadioSet

\item \code{phenoInfo<-}: Update the the given type of molecular data experiment
info in the RadioSet

\item \code{molecularProfiles}: Return the given type of molecular data from the RadioSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the RadioSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from
the RadioSet

\item \code{molecularProfilesSlot}: Get contents of molecularProfiles slot

\item \code{molecularProfilesSlot<-}: Update the molecular profiles slot of a RadioSet and
returns the updated copy

\item \code{featureInfo}: Return the feature info for the given molecular data

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{sensitivityInfo}: Return the radiation dose sensitivity experiment info

\item \code{sensitivityRaw}: Retrive the raw dose and viability data from an rSet

\item \code{sensitivityRaw<-}: Set the raw dose and viability data for an rSet and return
and updated copty

\item \code{sensitivitySlot}: Retrieve the contents of the sensitivity slot

\item \code{sensitivitySlot<-}: Set the raw dose and viability data for an rSet and return
and updated copty

\item \code{sensitivityInfo<-}: Update the sensitivity experiment info

\item \code{sensitivityProfiles}: Return the phenotypic data for the radiation dose
sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the radiation dose
sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the radiation dose
sensitivity

\item \code{sensitivityMeasures}: Returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{radiationTypes}: Return the names of the radiations used in the RadioSet

\item \code{radiationTypes<-}: Update the radiation names used in the dataset

\item \code{cellNames}: Return the cell names used in the dataset

\item \code{cellNames<-}: Update the cell names used in the dataset

\item \code{fNames}: Return the feature names used in the dataset

\item \code{fNames<-}: Set the feature names for a given molecular data type

\item \code{dateCreated}: Return the date the RadioSet was created

\item \code{datasetType}: Update the dataset type of an rSet and return a copy of
the updated object

\item \code{datasetType<-}: Update the dataset type of an rSet and return a copy of
the updated object

\item \code{name}: Return the name of the RadioSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{sensNumber}: Return a summary of available sensitivity experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments
}}

\section{Slots}{

\describe{
\item{\code{annotation}}{A \code{list} of annotation data about the RadioSet,
including the \code{$name} and the session information for how the object
was creating, detailing the exact versions of R and all the packages used}

\item{\code{molecularProfiles}}{A \code{list} containing 4 \code{SummarizedExperiment}
type object for holding data for RNA, DNA, SNP and Copy Number Variation
measurements respectively, with associated \code{fData} and \code{pData}
containing the row and column metadata}

\item{\code{cell}}{A \code{data.frame} containg the annotations for all the cell
lines profiled in the data set, across all data types}

\item{\code{radiation}}{A \code{data.frame} containg the annotations for all the
radiation treatment types used in the in the dataset, across all data types}

\item{\code{sensitivity}}{A \code{list} containing all the data for the sensitivity
experiments, including \code{$info}, a \code{data.frame} containing the
experimental info,\code{$raw} a 3D \code{array} containing raw data,
\code{$profiles}, a \code{data.frame} containing sensitivity profiles
statistics, and \code{$n}, a \code{data.frame} detailing the number of
experiments for each cell-radiation type pair}

\item{\code{perturbation}}{A \code{list} containting \code{$n}, a \code{data.frame}
summarizing the available perturbation data,}

\item{\code{curation}}{A \code{list} containing mappings for
\code{cell} and \code{tissue} names used in the data set to universal
identifiers used between different RadioSet objects}

\item{\code{datasetType}}{A \code{character} string of 'sensitivity',
'perturbation', or both detailing what type of data can be found in the
RadioSet, for proper processing of the data}
}}

\examples{
data(clevelandSmall)
cellInf <- cellInfo(clevelandSmall)

data(clevelandSmall)
cellInfo(clevelandSmall) <- cellInfo(clevelandSmall)

data(clevelandSmall)
phenoInf <- phenoInfo(clevelandSmall, mDataType="rna")

phenoInfo(clevelandSmall, mDataType="rna") <- phenoInfo(clevelandSmall,
  mDataType="rna")

data(clevelandSmall)
Cleveland_mProf <- molecularProfiles(clevelandSmall, "rna")

molecularProfiles(clevelandSmall, "rna") <-
  molecularProfiles(clevelandSmall, "rna")

data(clevelandSmall)
molecularProfilesSlot(clevelandSmall)

data(clevelandSmall_cSet)
molecularProfilesSlot(clevelandSmall_cSet) <- molecularProfilesSlot(clevelandSmall_cSet)

featInf <- featureInfo(clevelandSmall, "rna")

featureInfo(clevelandSmall, "rna") <- featureInfo(clevelandSmall, "rna")

sensInf<- sensitivityInfo(clevelandSmall)

sensitivityInfo(clevelandSmall) <- sensitivityInfo(clevelandSmall)

sensProf <- sensitivityProfiles(clevelandSmall)

sensitivityProfiles(clevelandSmall) <- sensitivityProfiles(clevelandSmall)

data(clevelandSmall)
sensMeas <- sensitivityMeasures(clevelandSmall)

data(clevelandSmall)
cellNames(clevelandSmall)

data(clevelandSmall)
cellNames(clevelandSmall) <- cellNames(clevelandSmall)

data(clevelandSmall)
fNames(clevelandSmall, "rna")

data(clevelandSmall)
fNames(clevelandSmall, 'rna') <- fNames(clevelandSmall, 'rna')

dateCreated(clevelandSmall)

name(clevelandSmall)

data(clevelandSmall)
pertNumber(clevelandSmall)

data(clevelandSmall)
sensNumber(clevelandSmall)

pertNumber(clevelandSmall) <- pertNumber(clevelandSmall)

sensNumber(clevelandSmall) <- sensNumber(clevelandSmall)

}
