% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet-class.R
\name{RadioSet}
\alias{RadioSet}
\title{RadioSet constructor}
\usage{
RadioSet(
  name,
  molecularProfiles = list(),
  cell = data.frame(),
  radiation = data.frame(),
  sensitivityInfo = data.frame(),
  sensitivityRaw = array(dim = c(0, 0, 0)),
  sensitivityProfiles = matrix(),
  sensitivityN = matrix(nrow = 0, ncol = 0),
  perturbationN = array(NA, dim = c(0, 0, 0)),
  curationCell = data.frame(),
  curationTissue = data.frame(),
  datasetType = c("sensitivity", "perturbation", "both"),
  verify = TRUE
)
}
\value{
An object of class RadioSet
}
\description{
A constructor that simplifies the process of creating RadioSets, as well
as creates empty objects for data not provided to the constructor. Only
objects returned by this constructor are expected to work with the RadioSet
methods. For a much more detailed instruction on creating RadioSets, please
see the "CreatingRadioSet" vignette.
}
