% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\name{RadioSig}
\alias{RadioSig}
\title{Radiation Signature Class Constructor}
\usage{
RadioSig(
  Data = array(NA, dim = c(0, 0, 0)),
  RSetName = "",
  DateCreated = date(),
  SigType = "sensitivity",
  SessionInfo = sessionInfo(),
  Call = "No Call Recorded"
)
}
\arguments{
\item{Data}{The data}

\item{RSetName}{The name of the pSet}

\item{DateCreated}{The date the object was created}

\item{SigType}{The type of sensitivyt signature}

\item{SessionInfo}{The package version used to generate the object}

\item{Call}{The calls for sensitivity vs not}
}
\value{
A \code{RadioSig} object
}
\description{
A documented constructor to provide user friendly interface to .RadioSig
}
