% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankGeneRadSensitivity.R
\name{rankGeneRadSensitivity}
\alias{rankGeneRadSensitivity}
\title{Rank genes based on radiation effect in the Connectivity Map}
\usage{
rankGeneRadSensitivity(
  data,
  drugpheno,
  type,
  batch,
  single.type = FALSE,
  standardize = "SD",
  nthread = 1,
  verbose = FALSE
)
}
\arguments{
\item{data}{gene expression data matrix}

\item{drugpheno}{sensititivity values fo thr drug of interest}

\item{type}{cell or tissue type for each experiment}

\item{batch}{experiment batches}

\item{single.type}{Should the statitsics be computed for each cell/tissue
type separately?}

\item{standardize}{How to standardize the data? Currently only supports "SD"}

\item{nthread}{number of parallel threads (bound to the maximum number of cores available)}

\item{verbose}{Should details of function operation be printed to console?}
}
\value{
A \code{list} of data.frames with the statistics for each gene, for
  each type
}
\description{
Rank genes based on radiation effect in the Connectivity Map
}
