% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{rSetName}
\alias{rSetName}
\title{rSetName Generic}
\usage{
rSetName(rSet)
}
\arguments{
\item{rSet}{A \code{RadioSet}}
}
\value{
The name of the RadioSet
}
\description{
A generic for the rSetName method
}
\examples{
data(CCLEsmall)
rSetName(CCLEsmall)

}
