% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivityRaw-methods.R
\name{sensitivityRaw<-}
\alias{sensitivityRaw<-}
\title{sensitivityRaw<- Replacement Generic}
\usage{
sensitivityRaw(object, ...) <- value
}
\arguments{
\item{object}{A \code{RadioSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}

\item{value}{A \code{array} containing the raw dose and viability data for the
rSet}
}
\value{
A copy of the \code{RadioSet} containing the updated sensitivty data
}
\description{
sensitivityRaw<- Replacement Generic
}
\examples{
data(clevelandSmall)
sensitivityRaw(clevelandSmall) <- sensitivityRaw(clevelandSmall)

}
