% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSetClass.R
\name{sensitivitySlot<-}
\alias{sensitivitySlot<-}
\title{sensitivitySlot<- Replacement Generic}
\usage{
sensitivitySlot(object, ...) <- value
}
\arguments{
\item{object}{A \code{RadioSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}

\item{value}{A \code{list} of new sensitivity slot data for the rSet}
}
\value{
A copy of the \code{RadioSet} containing the updated sensitivty slot
}
\description{
sensitivitySlot<- Replacement Generic
}
\examples{
data(clevelandSmall)
sensitivitySlot(clevelandSmall) <- sensitivitySlot(clevelandSmall)

}
