% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivitySlot-methods.R
\name{sensitivitySlot}
\alias{sensitivitySlot}
\title{sensitivitySlot Generic}
\usage{
sensitivitySlot(object, ...)
}
\arguments{
\item{object}{A \code{RadioSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}
}
\value{
A \code{list} of the sensitivity slot contents
}
\description{
sensitivitySlot Generic
}
\examples{
data(clevelandSmall)
sensitivitySlot(clevelandSmall)

}
