% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\docType{methods}
\name{show,RadioSet-method}
\alias{show,RadioSet-method}
\title{Show a PharamcoSet}
\usage{
\S4method{show}{RadioSet}(object)
}
\arguments{
\item{object}{\code{RadioSet}}
}
\value{
Prints the RadioSet object to the output stream, and returns invisible NULL.
}
\description{
Show a PharamcoSet
}
\examples{
data(CCLEsmall)
CCLEsmall

}
