% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\docType{methods}
\name{[,RadioSet-method}
\alias{[,RadioSet-method}
\title{`[`}
\usage{
\S4method{[}{RadioSet}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{RSet}

\item{i}{Cell lines to keep in RSet}

\item{j}{Drugs to keep in RSet}

\item{...}{further arguments}

\item{drop}{A boolean flag of whether to drop single dimensions or not}
}
\value{
Returns the subsetted RSet
}
\description{
`[`
}
