% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject-methods.R
\name{updateObject,RadioSet-method}
\alias{updateObject,RadioSet-method}
\title{Update the RadioSet class after changes in it struture or API}
\usage{
\S4method{updateObject}{RadioSet}(object)
}
\arguments{
\item{object}{A \code{RadioSet} object to update the class structure for.}
}
\value{
\code{RadioSet} with update class structure.
}
\description{
Update the RadioSet class after changes in it struture or API
}
