% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utlities.R
\name{.convertRsetMolecularProfilesToSE}
\alias{.convertRsetMolecularProfilesToSE}
\title{rSet molecularProfiles from ESets to SEs}
\usage{
.convertRsetMolecularProfilesToSE(rSet)
}
\arguments{
\item{rSet}{\code{S4} A RadioSet containing molecular data in ExpressionSets}
}
\value{
\code{S4} A RadioSet containing molecular data in a SummarizedExperiments
}
\description{
Converts all ExpressionSet objects within the molecularProfiles slot of a
  RadioSet to SummarizedExperiments
}
\keyword{internal}
