% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationTypes<-}
\alias{radiationTypes<-}
\title{radiationTypes<- Generic}
\usage{
radiationTypes(object) <- value
}
\value{
Updated \code{RadioSet}
}
\description{
A generic for the radiationTypes replacement method
}
\examples{
data(Cleveland_small)
radiationTypes(Cleveland_small) <- radiationTypes(Cleveland_small)

}
