% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationTypes}
\alias{radiationTypes}
\title{radiationTypes Generic}
\usage{
radiationTypes(rSet)
}
\value{
A vector of the radiation names used in the RadioSet
}
\description{
A generic for the radiationTypes method
}
\examples{
data(Cleveland_small)
radType <- radiationTypes(Cleveland_small)
radType[1:10]

}
