% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet-class.R
\name{RadioSet}
\alias{RadioSet}
\title{RadioSet constructor}
\usage{
RadioSet(
  name,
  molecularProfiles = list(),
  cell = data.frame(),
  radiation = data.frame(),
  sensitivityInfo = data.frame(),
  sensitivityRaw = array(dim = c(0, 0, 0)),
  sensitivityProfiles = matrix(),
  sensitivityN = matrix(nrow = 0, ncol = 0),
  perturbationN = array(NA, dim = c(0, 0, 0)),
  curationCell = data.frame(),
  curationTissue = data.frame(),
  datasetType = c("sensitivity", "perturbation", "both"),
  verify = TRUE
)
}
\arguments{
\item{name}{A \code{character} string detailing the name of the dataset}

\item{molecularProfiles}{A \code{list} of SummarizedExperiment objects containing
molecular profiles for each molecular data type.}

\item{sensitivityInfo}{A \code{data.frame} containing the information for the
sensitivity experiments. Must contain a 'sampleid' column with unique
identifiers to each sample, matching the \code{sample} object and a 'treatmentid'
columns with unique indenifiers for each treatment, matching the \code{treatment}
object.}

\item{sensitivityRaw}{A 3 Dimensional \code{array} contaning the raw drug
dose response data for the sensitivity experiments}

\item{sensitivityProfiles}{\code{data.frame} containing drug sensitivity profile
statistics such as IC50 and AUC}

\item{sensitivityN, perturbationN}{A \code{data.frame} summarizing the
available sensitivity/perturbation data}

\item{datasetType}{A \code{character(1)} string of 'sensitivity',
'preturbation', or 'both' detailing what type of data can be found in the
\code{CoreSet}, for proper processing of the data}

\item{verify}{\code{logical(1)}Should the function verify the CoreSet and
print out any errors it finds after construction?}
}
\value{
An object of class RadioSet
}
\description{
A constructor that simplifies the process of creating RadioSets, as well
as creates empty objects for data not provided to the constructor. Only
objects returned by this constructor are expected to work with the RadioSet
methods. For a much more detailed instruction on creating RadioSets, please
see the "CreatingRadioSet" vignette.
}
