% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\docType{methods}
\name{show,RadioSig-method}
\alias{show,RadioSig-method}
\title{Show RadioGx Signatures}
\usage{
\S4method{show}{RadioSig}(object)
}
\arguments{
\item{object}{\code{RadioSig}}
}
\value{
Prints the RadioGx Signatures object to the output stream, and returns invisible NULL.
}
\description{
Show RadioGx Signatures
}
\examples{
data(GDSCsmall)
drug.sensitivity <- drugSensitivitySig(GDSCsmall, mDataType="rna",
             nthread=1, features = fNames(GDSCsmall, "rna")[1])
drug.sensitivity

}
